requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
plotly
Pillow
pyyaml
catboost
pycaret
mlflow
scikit-plot
hydra-core